refactor(auto): YAML化任务编排并统一视觉抓取与ArmMove链路
- 任务编排改为配置驱动
  - 移除 task_builder 硬编码 TaskPlan 生成逻辑
  - 新增 task_catalog / yaml_parser / presets.yaml / tasks/*.yaml
  - AUTO_INIT / AUTO_GRAB / AUTO_STORE / AUTO_GET / FIXED_GRAB / TEST_* 迁移至 YAML
  - 扩展 protocol 支持 TEST_NOMAL，补充 auto.md 规则说明

- 机械臂动作接口统一为 ArmMove
  - 使用 ArmMove.action 替换 Move.action，清理旧消息定义
  - arm_solve_client / arm_solve_server / executor / straight_planner 全链路切换
  - 统一关节表与 target_joints 为 double 精度
  - step_executor 输出绑定改为 double_array

- 视觉抓取链路重构
  - detect_node 拆分为 segmentation / pointcloud_processor / pose_estimator / pose_from_axis_node
  - 新增抓取位姿服务、中心点/轴向发布、IoU 跟踪、圆柱拟合与 ICP 姿态估计参数
  - 更新 vision launch / yaml / 模型与 CAD 资源安装
  - vision_detect_client 保持 service 异步轮询模型

- 脚本与运行入口整理
  - rerun.sh 改为 managed_processes 统一管理
  - 切换 vision.launch.py 入口
  - term_launch 增加无 gnome-terminal 的降级支持
  - param_set.sh 补充 ROS 环境加载与参数校验

+++ src/capabilities/arm/arm_solve_client/include/arm_solve_client/arm_types.hpp
@@ -16,7 +16,7 @@ enum class PlanOption : uint8_t {
 
 struct ArmMoveSpec {
   engineer_interfaces::msg::Pose pose{};
-  std::array<float, 6> joints{{0.f, 0.f, 0.f, 0.f, 0.f, 0.f}};
+  std::array<double, 6> joints{{0.0, 0.0, 0.0, 0.0, 0.0, 0.0}};
   geometry_msgs::msg::Vector3 vector{};
 
   PlanOption plan_option{PlanOption::NORMAL};
